requirements.txt
fastapi==0.115.11
uvicorn[standard]==0.34.0
jinja2==3.1.6
python-multipart==0.0.20
sqlalchemy==2.0.39
psycopg[binary]==3.2.6
pydantic==2.11.2
pydantic-settings==2.8.1
redis==5.2.1
rq==2.1.0
beautifulsoup4==4.13.3
lxml==5.3.1
httpx==0.28.1
cryptography==44.0.2
google-api-python-client==2.162.0
google-auth==2.38.0
google-auth-oauthlib==1.2.1

pytest==8.3.5
